Added a grounded variable, reorganized code and betterified wall jumping.
With proper telling of if the object is grounded we can change the mask
during a jump flip and not have it bug out in the ceiling.

Code that was repeated in multiple files were merged into external functions.

Walljumping is now 'w' and pushes the player away from the wall.


DERPS:

Forgot to reverse lmb and rmb back.
Possibly created more bugs.